+++ drivers/flash/flash_stm32.h
@@ -10,6 +10,7 @@
 #define ZEPHYR_DRIVERS_FLASH_FLASH_STM32_H_
 
 #include <zephyr/drivers/flash.h>
+#include "stm32_hsem.h"
 
 #if DT_NODE_HAS_PROP(DT_INST(0, st_stm32_flash_controller), clocks) || \
 	DT_NODE_HAS_PROP(DT_INST(0, st_stm32h7_flash_controller), clocks)
@@ -270,6 +271,40 @@ static inline bool flash_stm32_range_exists(const struct device *dev,
 		 flash_get_page_info_by_offs(dev, offset + len - 1, &info));
 }
 #endif	/* CONFIG_FLASH_PAGE_LAYOUT */
+
+
+#if defined(CONFIG_MULTITHREADING) || defined(CONFIG_STM32H7_DUAL_CORE)
+/*
+ * This is named flash_stm32_sem_take instead of flash_stm32_lock (and
+ * similarly for flash_stm32_sem_give) to avoid confusion with locking
+ * actual flash pages.
+ */
+
+static inline void _flash_stm32_sem_take(const struct device *dev)
+{
+	k_sem_take(&FLASH_STM32_PRIV(dev)->sem, K_FOREVER);
+	z_stm32_hsem_lock(CFG_HW_FLASH_SEMID, HSEM_LOCK_WAIT_FOREVER);
+}
+
+static inline void _flash_stm32_sem_give(const struct device *dev)
+{
+	z_stm32_hsem_unlock(CFG_HW_FLASH_SEMID);
+	k_sem_give(&FLASH_STM32_PRIV(dev)->sem);
+}
+
+#define flash_stm32_sem_init(dev) k_sem_init(&FLASH_STM32_PRIV(dev)->sem, 1, 1)
+#define flash_stm32_sem_take(dev) _flash_stm32_sem_take(dev)
+#define flash_stm32_sem_give(dev) _flash_stm32_sem_give(dev)
+#else
+#define flash_stm32_sem_init(dev)
+#define flash_stm32_sem_take(dev)
+#define flash_stm32_sem_give(dev)
+#endif /* CONFIG_MULTITHREADING */
+
+#ifdef CONFIG_FLASH_EX_OP_ENABLED
+int flash_stm32_ex_op(const struct device *dev, uint16_t code,
+			     const uintptr_t in, void *out);
+#endif /* CONFIG_FLASH_EX_OP_ENABLED */
 
 static inline bool flash_stm32_valid_write(off_t offset, uint32_t len)
 {
